refactor: use next-image
