+++ src/views/product-landing/components/themed-image/themed-image.module.css
@@ -1,0 +1,12 @@
+.root {
+	align-items: center;
+	display: flex;
+	width: 100%;
+	height: 200px;
+	position: relative;
+}
+
+.image {
+	object-position: center;
+	object-fit: contain;
+}
