--- src/views/product-landing/components/themed-image/types.ts
@@ -1,7 +1,0 @@
-export interface ThemedImageProps {
-	src: {
-		dark: string
-		light: string
-	}
-	alt: string
-}
